@@ -1,7 +1,7 @@
 {
   "name": "functions",
   "scripts": {
-    "build": "tsc -p .",
+    "build": "eslint ./src && tsc -p .",
     "serve": "npm run build && firebase serve --only functions",
     "shell": "npm run build && firebase emulators:start --only functions",
     "start": "npm run shell",
@@ -27,19 +27,21 @@
   ],
   "dependencies": {
     "@iota/area-codes": "^0.1.1",
-    "@iota/client": "^2.1.1",
-    "axios": "^0.21.1",
+    "@iota/client": "^2.2.1",
+    "axios": "^0.24.0",
     "bluebird": "^3.7.2",
     "cors": "^2.8.5",
-    "firebase-admin": "^9.5.0",
-    "firebase-functions": "^3.13.1",
+    "firebase-admin": "^10.0.0",
+    "firebase-functions": "^3.16.0",
     "mailgun-js": "^0.22.0",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@firebase/app-types": "^0.6.1",
-    "tslint": "^6.1.3",
-    "typescript": "^3.8.3"
+    "@firebase/app-types": "^0.7.0",
+    "@typescript-eslint/eslint-plugin": "^5.3.0",
+    "@typescript-eslint/parser": "^5.3.0",
+    "eslint": "^8.1.0",
+    "typescript": "^4.4.4"
   },
   "engines": {
     "node": "14"
